+++ paracode-repl/src/main.rs
@@ -1,3 +1,5 @@
+pub mod repl;
+
 fn main() {
     println!("Hello, world!");
 }