+++ paracode/src/interpreter.rs
@@ -1,3 +1,6 @@
+pub mod basic_wrapper;
 pub mod basic_value;
+pub mod basic_object;
 
+pub mod typing;
 pub mod std;