+++ paracode/src/interpreter/basic_object.rs
@@ -1,0 +1,9 @@
+#[derive(Debug)]
+pub struct BasicObject {
+}
+impl BasicObject {
+    pub fn get_detailed_string(&self) -> String {
+        // TODO: implement
+        return "BasicObject".to_string();
+    }
+}