+++ paracode/src/interpreter/basic_value.rs
@@ -1,58 +1,127 @@
 use std::fmt;
-// use std::collections::HashMap;
+use std::fmt::Debug;
+use std::collections::HashMap;
 
-use crate::interpreter::std::float32::Float32;
+//use crate::interpreter::basic_object::BasicObject;
+//use crate::interpreter::typing::basic_type::BasicType;
 
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub enum BasicValue {
-    Int(i32),
-    // Use of a custom float wrapper instead of the primitive type is necesary for hash support, and may be removed later
-    Float(Float32),
-    Bool(bool),
-    String(String),
-    // TODO: Either get dictionaries to work or remove the below line
-    // Dictionary(HashMap<BasicValue, BasicValue>),
-    Array(Vec<BasicValue>),
-    Null,
-    // TODO: Add BasicObject and BasicType
-    // TODO: Add an extension point for user-defined types
+pub trait BasicValue: Debug {
+    fn compare_value(&self, other: Box<dyn BasicValue>) -> bool {
+        return self.get_detailed_string() == other.get_detailed_string();
+    }
+
+    // lookup_type
+
+    fn is_null(&self) -> bool {
+        return false;
+    }
+
+    fn is_object(&self) -> bool {
+        return false;
+    }
+
+    fn clone(&self) -> Self where Self: Sized;
+
+    fn to_string(&self) -> String;
+    fn get_detailed_string(&self) -> String;
 }
-impl BasicValue {
-    pub fn is_null(&self) -> bool {
-        return self == &BasicValue::Null;
+
+impl fmt::Display for dyn BasicValue {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        return write!(f, "{}", self.to_string());
     }
 }
 
-impl fmt::Display for BasicValue {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::Int(val) => write!(f, "{}", val),
-            Self::Float(val) => write!(f, "{}", val),
-            Self::Bool(val) => write!(f, "{}", val),
-            Self::String(val) => write!(f, "{}", val),
-            // Self::Dictionary(val) => write!(f, "{:?}", val),
-            Self::Array(val) => write!(f, "{:?}", val),
-            Self::Null => write!(f, "Null"),
+#[macro_export]
+macro_rules! impl_basicvalue {
+    ( $name:ident $(< $( $lt:tt $( : $clt:tt $(+ $dlt:tt )* )? ),+ >)? ) => {
+        impl $(< $( $lt $( : $clt $(+ $dlt )* )? ),+ >)? BasicValue for $name $(< $( $lt ),+ >)? {
+            fn is_null(&self) -> bool {
+                return stringify!($name) == "Null";
+            }
+            
+            fn is_object(&self) -> bool {
+                return stringify!($name) == "Empty";
+            }
+            
+            
+            fn clone(&self) -> Self {
+                return Clone::clone(self);
+            }
+    
+            fn to_string(&self) -> String {
+                return format!("{}", self);
+            }
+
+            fn get_detailed_string(&self) -> String {
+                return format!("Type: {}, Value: {:?}", stringify!($name), self);
+            }
         }
+    };
+}
+
+impl_basicvalue!(i32);
+impl_basicvalue!(f32);
+impl_basicvalue!(bool);
+impl_basicvalue!(String);
+
+impl<K, V, S> BasicValue for HashMap<K, V, S> where K: Clone + Debug, V: Clone + Debug, S: Clone + Debug {
+    fn clone(&self) -> Self {
+        return Clone::clone(self);
+    }
+
+    fn to_string(&self) -> String {
+        return format!("{:?}", self);
+    }
+
+    fn get_detailed_string(&self) -> String {
+        return format!("Type: {}, Value: {:?}", stringify!($typename), self);
     }
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn custom_float_addition() {
-      assert_eq!(Float32::new(1.4) + 2.1, Float32::new(1.4 + 2.1));
-    }
-
-    #[test]
-    fn null_is_null() {
-      assert_eq!(BasicValue::Null.is_null(), true);
-    }
-
-    #[test]
-    fn is_null() {
-      assert_eq!(BasicValue::Int(1).is_null(), false);
+// Rust doesn't have a null type, so create a fake one.
+#[derive(Debug, Clone)]
+pub struct Null;
+impl Null {
+    pub fn new() -> Self {
+        return Self {};
     }
 }
+impl fmt::Display for Null {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        return write!(f, "Null");
+    }
+}
+impl PartialEq<dyn BasicValue> for Null {
+    fn eq(&self, _rhs: &(dyn BasicValue + 'static)) -> bool {
+        return self.is_null() && _rhs.is_null();
+    }
+
+    fn ne(&self, _rhs: &(dyn BasicValue + 'static)) -> bool {
+        return !(self == _rhs);
+    }
+}
+impl_basicvalue!(Null);
+
+#[derive(Debug, Clone)]
+pub struct Empty;
+impl Empty {
+    pub fn new() -> Self {
+        return Self {};
+    }
+}
+impl fmt::Display for Empty {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        return write!(f, "Empty");
+    }
+}
+impl PartialEq<dyn BasicValue> for Empty {
+    fn eq(&self, _rhs: &(dyn BasicValue + 'static)) -> bool {
+        return self.is_object() && _rhs.is_object();
+    }
+
+    fn ne(&self, _rhs: &(dyn BasicValue + 'static)) -> bool {
+        return !(self == _rhs);
+    }
+}
+impl_basicvalue!(Empty);