+++ paracode/src/interpreter/typing/basic_type.rs
@@ -1,0 +1,9 @@
+#[derive(Debug)]
+pub struct BasicType {
+}
+impl BasicType {
+    pub fn get_detailed_string(&self) -> String {
+        // TODO: implement
+        return "BasicType".to_string();
+    }
+}